+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Development version
 
-_No changes yet._
+- Make the CLI an optional (but default) feature (#62). This speeds up the build when using Spreet as a Rust library (see [README](README.md#using-spreet-as-a-rust-library))
 
 ## v0.8.0 (2023-06-15)