+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Update [resvg](https://crates.io/crates/resvg) dependency to v0.35
 - Update [clap](https://crates.io/crates/clap) dependency to v4.4
 - Remove [Rayon](https://crates.io/crates/rayon) dependency. This means the Spreet CLI no longer parses SVGs in parallel, but that was a fun-but-unnecessary optimisation in the first place that generally saved only a handful of milliseconds
+- **Deprecated**: `spreet::sprite::generate_pixmap_from_svg()` has been deprecated and will be removed in a future version. Use `spreet::sprite::Spreet::pixmap()` instead
 
 ## v0.8.0 (2023-06-15)